after review: changes searchmanager, remove and create input in searchmanager
