--- src/app/InputSearch.js
@@ -1,9 +1,0 @@
-export class InputSearch {
-
-    createInput(){
-        let input = document.createElement('input');
-        input.id = 'autoComplete';
-        input.type = 'search';
-        document.querySelector('#zbiorkom-app').append(input);
-    }
-}
