+++ test/SearchManager.spec.js
@@ -1,0 +1,12 @@
+//import {SearchManager} from '../src/app/SearchManager';
+import '@testing-library/jest-dom/extend-expect';
+
+//przy imporcie searchmanagera jest błąd 'Missing class properties transform'
+describe('Tests input exists and values', () => {
+    test('nie umiem pisać testów', () =>{
+        //const searchManager = new SearchManager();
+        expect(true).toBe(true);
+        expect(true).toBeTruthy();
+        expect(true).not.toBeFalsy();
+    })
+})
